prepare hooks for oob navigation
