+++ src/components/search-header/second-row.ts
@@ -1,0 +1,106 @@
+import * as list from "@app/ds/list";
+import { fakeRovingTabIndex, findFirstFakeFocusable } from "@app/dom";
+import { setupCustomSelect } from "./fake-select";
+
+export function setupSecondRow(root: HTMLElement): {
+  teardown: Teardown;
+  allItems: HTMLElement[];
+} {
+  const selects = Array.from(
+    root.querySelectorAll<HTMLSelectElement>("select")
+  );
+  const teardown: Teardown[] = [];
+
+  // Create the custom select
+  teardown.push(...selects.map((select) => setupCustomSelect(select)));
+
+  const genreCustomSelect = root.querySelector<HTMLElement>(
+    '.fgn-custom-select[aria-label="Genre"]'
+  );
+  const releaseYearCustomSelect = root.querySelector<HTMLElement>(
+    '.fgn-custom-select[aria-label="Release year"]'
+  );
+  const systemCustomSelect = root.querySelector<HTMLElement>(
+    '.fgn-custom-select[aria-label="System"]'
+  );
+  const rankedCustomSelect = root.querySelector<HTMLElement>(
+    '.fgn-custom-select[aria-label="Ranked"]'
+  );
+  if (
+    !genreCustomSelect ||
+    !releaseYearCustomSelect ||
+    !systemCustomSelect ||
+    !rankedCustomSelect
+  ) {
+    throw new Error(
+      "Could not initialize filters, looks like something is missing"
+    );
+  }
+  return {
+    teardown: () => {
+      teardown.forEach((v) => v());
+    },
+    allItems: [
+      genreCustomSelect,
+      releaseYearCustomSelect,
+      systemCustomSelect,
+      rankedCustomSelect,
+    ],
+  };
+}
+
+export function setupSecondRowListeners(
+  firstRowItems: HTMLElement[],
+  secondRowItems: HTMLElement[],
+  thirdRowItems: HTMLElement[],
+  onHorizontalOOB: (direction: "START" | "END") => void,
+  onVerticalOOB: (direction: "START" | "END") => void
+) {
+  function onKeydown(currEl: HTMLElement, e: KeyboardEvent) {
+    const keyPressed = e.key;
+
+    if (keyPressed === "ArrowLeft") {
+      e.preventDefault();
+
+      const next = list.prev(secondRowItems, currEl);
+      if (next.status === "OK") {
+        fakeRovingTabIndex(currEl, next.value);
+        next.value.focus();
+      } else {
+        onHorizontalOOB("START");
+      }
+    } else if (keyPressed === "ArrowRight") {
+      e.preventDefault();
+
+      const next = list.next(secondRowItems, currEl);
+      if (next.status === "OK") {
+        fakeRovingTabIndex(currEl, next.value);
+        next.value.focus();
+      }
+    } else if (keyPressed === "ArrowUp") {
+      e.preventDefault();
+      findFirstFakeFocusable(firstRowItems)?.focus();
+    } else if (keyPressed === "ArrowDown") {
+      e.preventDefault();
+      // It may be possible for a third row for to not exist
+      const firstFocusable = findFirstFakeFocusable(thirdRowItems);
+      if (firstFocusable) {
+        firstFocusable.focus();
+      } else {
+        onVerticalOOB("END");
+      }
+    } else if (keyPressed === "Enter") {
+      currEl.click();
+    }
+  }
+
+  secondRowItems.forEach((el) => {
+    el.addEventListener("keydown", onKeydown.bind(null, el));
+  });
+
+  return () => {
+    secondRowItems.forEach((el) => {
+      el.addEventListener("keydown", onKeydown.bind(null, el));
+    });
+  };
+}
