+++ src/types/allyjs.d.ts
@@ -1,3 +1,4 @@
 declare module "ally.js/src/query/tabsequence";
 declare module "ally.js/src/query/focusable";
 declare module "ally.js/src/is/visible";
+declare module "ally.js/src/is/focusable";
